+++ webEdition/liveUpdate/sqldumps/tblversionslog.sql
@@ -1,6 +1,6 @@
 CREATE TABLE ###TBLPREFIX###tblversionslog (
-  `ID` bigint(20) unsigned NOT NULL auto_increment,
-  `timestamp` int(11) unsigned NOT NULL,
+  `ID` int(11) unsigned NOT NULL auto_increment,
+  `timestamp` int(10) unsigned NOT NULL,
   `action` int(11) unsigned NOT NULL,
   `userID` int(11) unsigned NOT NULL,
   `data` longtext NOT NULL,
